+++ AOC2019/IntCode/IntCodeInputHelper.cs
@@ -1,18 +1,18 @@
 ﻿namespace AOC2019.IntCode
 {
-    internal class IntCodeInputHelper : InputHelper<int[]>
+    internal class IntCodeInputHelper : InputHelper<long[]>
     {
         public IntCodeInputHelper(string fileName) : base(fileName)
         {
         }
 
-        public override int[] Parse()
+        public override long[] Parse()
         {
-            var output = new int[0];
+            var output = new long[0];
             using (var sr = new StreamReader(InputPath))
             {
                 var ints = sr.ReadLine()!.Split(',');
-                output = Array.ConvertAll(ints, int.Parse);
+                output = Array.ConvertAll(ints, long.Parse);
             }
             return output;
         }
